+++ src/components/App.js
@@ -7,13 +7,13 @@ var discoveryUrl = 'https://www.googleapis.com/discovery/v1/apis/drive/v3/rest';
 const API_KEY = config.web.api_key
 const CLIENT_ID = config.web.client_id
 
+let userId = 1
+
 class App extends Component {
   constructor() {
     super();
     this.state = {
-      name: '',
-      googleAuth: '',
-      userFiles: [],
+      userList: []
     }
   }
   componentDidMount() {
@@ -35,26 +35,7 @@ class App extends Component {
         'scope': SCOPE,
         'discoveryDocs': [discoveryUrl]
       }).then(() => {
-        this.setState({
-          googleAuth: window.gapi.auth2.getAuthInstance()
-        })
-        this.state.googleAuth.isSignedIn.listen(this.updateSigninStatus);
-        console.log(window.gapi)
-        // Setting parameters for list
-        console.log(this);
-        window.gapi.client.drive.files.list({
-          fields: 'files(id, name, mimeType, starred, iconLink, shared)'
-        })
-          .then(response => {
-            // Handle the results here (response.result has the parsed body).
-            console.log(this)
-            this.setState({
-              userFiles: response.result.files
-            });
-            console.log("Response", response);
-            console.log(response.result.files);
-          },
-            function (err) { console.error("Execute error", err); });
+        //this.state.googleAuth.isSignedIn.listen(this.updateSigninStatus);
         document.getElementById('signin-btn').addEventListener('click', this.signInFunction);
         document.getElementById('signout-btn').addEventListener('click', this.signOutFunction);
 
@@ -65,10 +46,15 @@ class App extends Component {
   }
 
   signInFunction = () => {
-    this.state.googleAuth.signIn();
-    this.updateSigninStatus();
+    window.gapi.auth2.getAuthInstance().signIn()
+    this.addUser()
+    const newUserIndex = this.state.userList.length - 1
+    this.updateFiles(newUserIndex, this.state.userList[newUserIndex].drive.files)
+    this.addUserInfo(newUserIndex, this.state.userList[newUserIndex].googleAuth.currentUser.get().rt)
+    
   }
 
+  //TODO: need to remove user from userlist when signing out
   signOutFunction = () => {
     this.state.googleAuth.signOut();
     this.updateSigninStatus();
@@ -78,6 +64,7 @@ class App extends Component {
     this.setSigninStatus();
   }
 
+  //sets the sign in display name - TODO: delete
   setSigninStatus = async () => {
     var user = this.state.googleAuth.currentUser.get();
     console.log(user)
@@ -95,16 +82,60 @@ class App extends Component {
       }
     }
   }
+
+  addUser = () => {
+    this.setState(prevState => {
+      return {
+        userList: [...prevState.userList, { id: userId++, drive: window.gapi.client.drive, googleAuth: window.gapi.auth2.getAuthInstance(), files: [], info: {} }]
+      }
+    })
+  }
+
+  //gets the files and stores them for user at given index
+  updateFiles = (index, files) => {
+    // check if right user
+    files.list({
+      fields: 'files(id, name, mimeType, starred, iconLink, shared)'
+    })
+      .then(response => {
+        // Handle the results here (response.result has the parsed body).
+        // update that user's files
+        this.setState(prevState => {
+          let newUserList = prevState.userList
+          newUserList[index].files = response.result.files
+          return {
+            userList: newUserList
+          }
+        })
+      },
+        function (err) { console.error("Execute error", err); });
+  }
+
+  addUserInfo = (index, info) => {
+    this.setState(prevState => {
+      let newUserList = prevState.userList
+      newUserList[index].info = info
+      return {
+        userList: newUserList
+      }
+    })
+  }
+
+
   render() {
     return (
       <div className="App">
         <div>UserName: <strong>{this.state.name}</strong></div>
         <button id="signin-btn">Sign In</button>
         <button id="signout-btn">Sign Out</button>
-        <User 
-          fileList = {this.state.userFiles}
-          name = {this.state.name}
-        />
+        {this.state.userList.map(user => {
+          return (
+            <User
+              name={user.info.Ad}
+              fileList={user.files}
+            />
+          )
+        })}
       </div>
     );
   }
